fixes issue on useLockend

+++ src/hooks/useLockEnd.ts
@@ -6,7 +6,7 @@ export const useLockEnd = () => {
   const [lockEndDate, setLockEndDate] = useState<Date>()
   const { userLockendDate } = useLockOverview()
   useEffect(() => {
-    const value = getUserLockEndDate(userLockendDate?.toString() ?? '')
+    const value = getUserLockEndDate((userLockendDate as bigint)?.toString() ?? '')
     setLockEndDate(value)
   }, [userLockendDate])
